requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.7.0
audioread==3.0.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
exceptiongroup==1.2.2
fastapi==0.115.6
ffmpy==0.4.0
filelock==3.16.1
fonttools==4.53.1
fsspec==2024.10.0
gradio==4.44.1
gradio_client==1.3.0
grpcio==1.66.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.26.5
idna==3.7
importlib_resources==6.4.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.8
numba==0.60.0
numpy==2.0.0
orjson==3.10.12
packaging==24.1
pandas==2.2.3
pillow==10.4.0
platformdirs==4.2.2
pooch==1.8.2
praat-parselmouth==0.4.3
protobuf==5.28.2
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-multipart==0.0.19
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
ruff==0.8.2
scikit-learn==1.5.1
scipy==1.13.1
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
soxr==0.4.0
starlette==0.41.3
sympy==1.13.3
TextGrid==1.6.1
threadpoolctl==3.5.0
tomlkit==0.12.0
tqdm==4.66.5
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.2
uvicorn==0.32.1
websockets==12.0
zipp==3.19.2
